refactor: use camelCase for schematype file names
