+++ lib/document.js
@@ -7,7 +7,7 @@
 const EventEmitter = require('events').EventEmitter;
 const InternalCache = require('./internal');
 const MongooseError = require('./error/index');
-const MixedSchema = require('./schema/Mixed');
+const MixedSchema = require('./schema/mixed');
 const ObjectExpectedError = require('./error/objectExpected');
 const ObjectParameterError = require('./error/objectParameter');
 const ParallelValidateError = require('./error/parallelValidate');
