+++ lib/helpers/model/discriminator.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const Mixed = require('../../schema/Mixed');
+const Mixed = require('../../schema/mixed');
 const applyBuiltinPlugins = require('../schema/applyBuiltinPlugins');
 const clone = require('../clone');
 const defineKey = require('../document/compile').defineKey;
