+++ lib/helpers/populate/getSchemaTypes.js
@@ -4,7 +4,7 @@
  * ignore
  */
 
-const Mixed = require('../../schema/Mixed');
+const Mixed = require('../../schema/mixed');
 const get = require('../get');
 const getDiscriminatorByValue = require('../discriminator/getDiscriminatorByValue');
 const leanPopulateMap = require('./leanPopulateMap');
