+++ lib/schema/DocumentArrayElement.js
@@ -6,7 +6,7 @@
 
 const MongooseError = require('../error/mongooseError');
 const SchemaType = require('../schemaType');
-const SchemaSubdocument = require('./Subdocument');
+const SchemaSubdocument = require('./subdocument');
 const getConstructor = require('../helpers/discriminator/getConstructor');
 
 /**
