+++ lib/schema/array.js
@@ -10,7 +10,7 @@ const MongooseError = require('../error/mongooseError');
 const SchemaArrayOptions = require('../options/SchemaArrayOptions');
 const SchemaType = require('../schemaType');
 const CastError = SchemaType.CastError;
-const Mixed = require('./Mixed');
+const Mixed = require('./mixed');
 const arrayDepth = require('../helpers/arrayDepth');
 const cast = require('../cast');
 const clone = require('../helpers/clone');
