+++ lib/schema/documentArray.js
@@ -7,7 +7,7 @@
 const CastError = require('../error/cast');
 const DocumentArrayElement = require('./DocumentArrayElement');
 const EventEmitter = require('events').EventEmitter;
-const SchemaArray = require('./Array');
+const SchemaArray = require('./array');
 const SchemaDocumentArrayOptions =
   require('../options/SchemaDocumentArrayOptions');
 const SchemaType = require('../schemaType');
