+++ lib/schema/index.js
@@ -5,20 +5,20 @@
 
 'use strict';
 
-exports.Array = require('./Array');
-exports.BigInt = require('./BigInt');
-exports.Boolean = require('./Boolean');
-exports.Buffer = require('./Buffer');
-exports.Date = require('./Date');
-exports.Decimal128 = exports.Decimal = require('./Decimal128');
-exports.DocumentArray = require('./DocumentArray');
-exports.Map = require('./Map');
-exports.Mixed = require('./Mixed');
-exports.Number = require('./Number');
-exports.ObjectId = require('./ObjectId');
-exports.String = require('./String');
-exports.Subdocument = require('./Subdocument');
-exports.UUID = require('./UUID');
+exports.Array = require('./array');
+exports.BigInt = require('./bigint');
+exports.Boolean = require('./boolean');
+exports.Buffer = require('./buffer');
+exports.Date = require('./date');
+exports.Decimal128 = exports.Decimal = require('./decimal128');
+exports.DocumentArray = require('./documentArray');
+exports.Map = require('./map');
+exports.Mixed = require('./mixed');
+exports.Number = require('./number');
+exports.ObjectId = require('./objectId');
+exports.String = require('./string');
+exports.Subdocument = require('./subdocument');
+exports.UUID = require('./uuid');
 
 // alias
