+++ lib/schema/map.js
