+++ lib/schema/objectId.js
@@ -278,7 +278,7 @@ function defaultId() {
 defaultId.$runBeforeSetters = true;
 
 function resetId(v) {
-  Document || (Document = require('./../document'));
+  Document || (Document = require('../document'));
 
   if (this instanceof Document) {
     if (v === void 0) {
