+++ lib/schema/operators/geospatial.js
@@ -16,7 +16,7 @@ exports.cast$near = cast$near;
 exports.cast$within = cast$within;
 
 function cast$near(val) {
-  const SchemaArray = require('../Array');
+  const SchemaArray = require('../array');
 
   if (Array.isArray(val)) {
     castArraysOfNumbers(val, this);
