+++ lib/schema/operators/helpers.js
@@ -4,7 +4,7 @@
  * Module requirements.
  */
 
-const SchemaNumber = require('../Number');
+const SchemaNumber = require('../number');
 
 /*!
  * ignore
