+++ lib/schema/string.js
