+++ lib/types/map.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const Mixed = require('../schema/Mixed');
+const Mixed = require('../schema/mixed');
 const MongooseError = require('../error/mongooseError');
 const clone = require('../helpers/clone');
 const deepEqual = require('../utils').deepEqual;
